Cached API routes/serverless function responses
